chore: adjust dom
