+++ packages/components/src/virtual-list/index.tsx
@@ -1,32 +1,7 @@
 import React, { useEffect, useRef } from 'react';
-import { ListRange, Virtuoso, VirtuosoHandle } from 'react-virtuoso';
+import { ScrollerProps, Virtuoso, VirtuosoHandle } from 'react-virtuoso';
 
-export type IVirtualListHandle = VirtuosoHandle;
-export type IVirtualListRange = ListRange;
-
-export interface IVirtualListProps {
-  /**
-   * List数据源
-   * @type {any[]}
-   */
-  data: any[];
-  /**
-   * 基础数据源渲染模板
-   * 默认传入参数为：(data, index) => {}
-   * data 为 this.props.data中的子项
-   * index 为当前下标
-   * @type {React.ComponentType<any>}
-   */
-  template: React.ComponentType<any>;
-  /**
-   * List外部样式名
-   * @type {string}
-   */
-  className?: string;
-  refSetter?: (virtuoso: IVirtualListHandle | null) => void;
-
-  onRangeChanged?: (range: IVirtualListRange) => void;
-}
+import { IVirtualListProps } from './types';
 
 export const VirtualList = ({
   data,
